requirements.txt
networkx==3.6.1
numpy==2.4.4
pygame-ce==2.5.7
